TASK:基本完成用户管理微服务功能

+++ src/main/java/com/primeton/maoxinjie/demo/service/impl/OrganizationServiceImpl.java
@@ -1,13 +1,20 @@
 package com.primeton.maoxinjie.demo.service.impl;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.github.pagehelper.PageHelper;
+import com.github.pagehelper.PageInfo;
+import com.primeton.maoxinjie.demo.constant.ResultCodeEnum;
 import com.primeton.maoxinjie.demo.dao.IOrganizationDao;
 import com.primeton.maoxinjie.demo.dao.IUserDao;
 import com.primeton.maoxinjie.demo.exception.BusiException;
 import com.primeton.maoxinjie.demo.model.OrganizationModel;
+import com.primeton.maoxinjie.demo.model.UserModel;
 import com.primeton.maoxinjie.demo.service.IOrganizationService;
 
 /**
@@ -20,6 +27,7 @@ import com.primeton.maoxinjie.demo.service.IOrganizationService;
  */
 @Service
 @JsonIgnoreProperties(ignoreUnknown=true)
+@Transactional
 public class OrganizationServiceImpl implements IOrganizationService {
 	
 	@Autowired
@@ -38,12 +46,12 @@ public class OrganizationServiceImpl implements IOrganizationService {
 	public int createOrganization(OrganizationModel organizationModel) throws Exception {
 		int num = 0;
 		if(organizationModel == null) {
-			throw new BusiException("所创建的组织机构信息为空!");
+			throw new BusiException(ResultCodeEnum.ORG_NULL_ERROR.getCode(),ResultCodeEnum.ORG_NULL_ERROR.getMessage());
 		}
 		if(organizationDao.getOrganizationByOrgName(organizationModel.getoName()) != null) {
-			throw new BusiException("该组织机构已存在!");
+			throw new BusiException(ResultCodeEnum.ORG_EXIST_ERROR.getCode(),ResultCodeEnum.ORG_EXIST_ERROR.getMessage());
 		}
-		num = organizationDao.insertOrganization(organizationModel);
+			num = organizationDao.insertOrganization(organizationModel);
 		return num;
 	}
 
@@ -70,11 +78,11 @@ public class OrganizationServiceImpl implements IOrganizationService {
 	public int modifyOrganization(OrganizationModel organizationModel) throws Exception {
 		int num = 0;
 		if(organizationModel == null) {
-			throw new BusiException("修改的组织机构信息为空!");
+			throw new BusiException(ResultCodeEnum.ORG_NULL_ERROR.getCode(),ResultCodeEnum.ORG_NULL_ERROR.getMessage());
 		}
 		OrganizationModel record = organizationDao.getOrganizationByID(organizationModel.getId());
 		if(record == null || record.getoName().equals(organizationModel.getoName())) {
-			throw new BusiException("用户名已存在不能进行修改");
+			throw new BusiException(ResultCodeEnum.ORG_EXIST_ERROR.getCode(),ResultCodeEnum.ORG_EXIST_ERROR.getMessage());
 		}
 			num = organizationDao.updateOrganization(organizationModel);
 		return num;
@@ -89,7 +97,6 @@ public class OrganizationServiceImpl implements IOrganizationService {
 	 */
 	@Override
 	public OrganizationModel getOrganizationByID(int id) throws Exception {
-		
 		return organizationDao.getOrganizationByID(id);
 	}
 
@@ -102,7 +109,37 @@ public class OrganizationServiceImpl implements IOrganizationService {
 	 */
 	@Override
 	public int removeOrganizationById(int id) throws Exception {
-		return organizationDao.deleteOrganizationById(id);
+		int num = 0;
+		//首先判断id对应的信息是否存在
+		OrganizationModel record = organizationDao.getOrganizationByID(id);
+		if(record == null) {
+			throw new BusiException(ResultCodeEnum.ID_ORG_EXIST_ERROR.getCode(),ResultCodeEnum.ID_ORG_EXIST_ERROR.getMessage());
+		}
+		//在判断该组织机构对应的有没有用户
+		List<UserModel> userList = userDao.queryUserByOID(id);
+		if(userList.size() > 0) {
+			throw new BusiException(ResultCodeEnum.ORG_USER_NOT_NULL_ERROR.getCode(), ResultCodeEnum.ORG_USER_NOT_NULL_ERROR.getMessage());
+		}
+			num = organizationDao.deleteOrganizationById(id);
+		return num;
+	}
+
+	/**
+	 * 
+	 * <p>Description: 组织机构的分页+搜索查询</p> 
+	 * @param pageNo
+	 * @param pageSize
+	 * @param organizationModel
+	 * @return
+	 * @throws Exception
+	 */
+	@Override
+	public PageInfo<OrganizationModel> queryOrgByPage(int pageNo, int pageSize, OrganizationModel organizationModel)
+			throws Exception {
+		//对于官方文档所说PageHelper方法调用后紧跟 MyBatis 查询方法，这就是安全的
+		PageHelper.startPage(pageNo, pageSize);
+		List<OrganizationModel> orgList = organizationDao.queryOrganizationByPage(organizationModel);
+	return new PageInfo<>(orgList);
 	}
 
 }
